Move analyze API implementation to separate module
